+++ ChordChartApp/src/main/java/com/oneandahalfasians/chordchartapp/data/entities/line/LyricLine.java
@@ -3,17 +3,17 @@ package com.oneandahalfasians.chordchartapp.data.entities.line;
 import java.util.ArrayList;
 import java.util.List;
 
-public class LyricLine {
-    private List<Lyric> lyricList = new ArrayList<>();
+public class LyricLine<LyricType> {
+    private List<LyricType> lyricList = new ArrayList<>();
 
     public LyricLine() {
     }
 
-    public List<Lyric> getLyricList() {
+    public List<LyricType> getLyricList() {
         return lyricList;
     }
 
-    public void setLyricList(List<Lyric> lyricList) {
+    public void setLyricList(List<LyricType> lyricList) {
         this.lyricList = lyricList;
     }
 }
